move emr_container hook
