+++ airflow/providers/amazon/aws/operators/emr.py
@@ -30,7 +30,7 @@ if sys.version_info >= (3, 8):
 else:
     from cached_property import cached_property
 
-from airflow.providers.amazon.aws.hooks.emr_containers import EMRContainerHook
+from airflow.providers.amazon.aws.hooks.emr import EmrContainerHook
 
 
 class EmrAddStepsOperator(BaseOperator):
@@ -177,9 +177,9 @@ class EmrContainerOperator(BaseOperator):
         self.job_id = None
 
     @cached_property
-    def hook(self) -> EMRContainerHook:
-        """Create and return an EMRContainerHook."""
-        return EMRContainerHook(
+    def hook(self) -> EmrContainerHook:
+        """Create and return an EmrContainerHook."""
+        return EmrContainerHook(
             self.aws_conn_id,
             virtual_cluster_id=self.virtual_cluster_id,
         )
@@ -196,13 +196,13 @@ class EmrContainerOperator(BaseOperator):
         )
         query_status = self.hook.poll_query_status(self.job_id, self.max_tries, self.poll_interval)
 
-        if query_status in EMRContainerHook.FAILURE_STATES:
+        if query_status in EmrContainerHook.FAILURE_STATES:
             error_message = self.hook.get_job_failure_reason(self.job_id)
             raise AirflowException(
                 f"EMR Containers job failed. Final state is {query_status}. "
                 f"query_execution_id is {self.job_id}. Error: {error_message}"
             )
-        elif not query_status or query_status in EMRContainerHook.INTERMEDIATE_STATES:
+        elif not query_status or query_status in EmrContainerHook.INTERMEDIATE_STATES:
             raise AirflowException(
                 f"Final state of EMR Containers job is {query_status}. "
                 f"Max tries of poll status exceeded, query_execution_id is {self.job_id}."
