+++ dev/provider_packages/prepare_provider_packages.py
@@ -2156,11 +2156,12 @@ KNOWN_DEPRECATED_DIRECT_IMPORTS: Set[str] = {
     "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.dms`.",
     'This module is deprecated. Please use `airflow.providers.amazon.aws.operators.emr`.',
     'This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.emr`.',
-    "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.redshift_cluster` "
-    "or `airflow.providers.amazon.aws.hooks.redshift_sql` as appropriate.",
-    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.redshift_sql` "
-    "or `airflow.providers.amazon.aws.operators.redshift_cluster` as appropriate.",
-    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.redshift_cluster`.",
+    'This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.emr`.',
+    'This module is deprecated. Please use `airflow.hooks.redshift_sql` '
+    'or `airflow.hooks.redshift_cluster` as appropriate.',
+    'This module is deprecated. Please use `airflow.operators.redshift_sql` or '
+    '`airflow.operators.redshift_cluster` as appropriate.',
+    'This module is deprecated. Please use `airflow.sensors.redshift_cluster`.',
 }
