+++ tests/deprecated_classes.py
@@ -252,6 +252,10 @@ HOOKS = [
         'airflow.contrib.hooks.aws_logs_hook.AwsLogsHook',
     ),
     (
+        'airflow.providers.amazon.aws.hooks.emr.EmrContainerHook',
+        'airflow.providers.amazon.aws.hooks.emr_containers.EMRContainerHook',
+    ),
+    (
         'airflow.providers.amazon.aws.hooks.emr.EmrHook',
         'airflow.contrib.hooks.emr_hook.EmrHook',
     ),
